+++ src/rules.js
@@ -1,0 +1,4 @@
+export default {
+	indent: 2,
+	eol: '\n'
+};