+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.0.1] - 07-07-2016
+### Added
+- Added #12, clinton
+- Added support boolean attributes
+- Added attributes list
+
+### Change
+- Change #13, change xo to eslint
+- Change #11, the rules should be as property
+- Update tests
+
+### Fixed
+- Fixed linter report error
+
 ## [0.0.0] - 01-07-2016
 ### Added
 - Init #1, project